chore: renaming and return types
