+++ packages/shared/lib/core/wallet/stores/selected-account-activities.store.ts
@@ -5,9 +5,10 @@ import { localize } from '@core/i18n'
 import { formatUnitBestMatch } from '@lib/units'
 import { getMonthYear, isValueInUnitRange, unitToValue } from '@lib/utils'
 
-import { Activity, ActivityFilter } from '..'
+import { Activity } from '../classes/activity.class'
 import { allAccountActivities } from './all-account-activities.store'
 import { isFilteredActivity } from '../utils/isFilteredActivity'
+import { ActivityFilter } from '../interfaces/filter.interface'
 
 export const selectedAccountActivities: Readable<Activity[]> = derived(
     [selectedAccount, allAccountActivities],
@@ -21,7 +22,7 @@ export const selectedAccountActivities: Readable<Activity[]> = derived(
 )
 
 export const activityFilter: Writable<ActivityFilter> = writable({
-    showHidden: { active: false, type: 'boolean', label: 'filters.activity.showHidden' },
+    showHidden: { active: false, type: 'boolean', label: 'filters.showHidden' },
 })
 export const activitySearchTerm: Writable<string> = writable('')
