+++ packages/shared/lib/core/wallet/utils/isFilteredActivity.ts
@@ -5,7 +5,7 @@ import { activityFilter } from '../stores'
 export function isFilteredActivity(activity: Activity): boolean {
     const filter = get(activityFilter)
 
-    if (!filter.showHidden.active && filter.showHidden.active && activity.isAssetHidden) {
+    if (!filter.showHidden.active && activity.isAssetHidden) {
         return true
     }
     return false
